// client-id/src/app/core/models/inspection-popup/ipt-vesslT.ts
export class IptVesslT{
    mode: string;
    athrzSn: number;
    athrzDetailSn: number;
    reqstQy: number;
    vesslNoStr: string;
    vesslNo: string;
    vesslCnt: string;
    hVesslCnt: number;
    reqstEntrpsNm: string;
    page: number;

    constructor(){
      this.mode = null;
      this.athrzSn = 0;
      this.athrzDetailSn = 0;
      this.reqstQy = 0;
      this.reqstEntrpsNm = '';
      this.vesslNoStr = null;
      this.vesslNo = null;
      this.hVesslCnt = null;
      this.hVesslCnt = null;
      this.page = 0;
    }
  }